So, I'm very surly that I AM A FOOL!!!
In front-end design, the wrong ID of button is not seen, and then it is mistakenly thought to be caused by repeated variable names. Later I thought, no!! The variables under the two classes will not conflict. So, I AM FOOLISH!!!

OK, anyway, this time we create the activity_student_sign_up.xml and activity_teacher_sign_up.xml so that we could register different accounts from different webs. And in the next step, we could separate the database task into two parts. One is for the student and another for the teacher.